refactor: move shared error messages to sdk
