+++ packages/adapter-ccxt/src/validation/StepConfigValidator.ts
@@ -1,5 +1,9 @@
 import { injectable } from "tsyringe";
-import { IValidator } from "@mate/sdk/src/interfaces/IValidator";
+import {
+   IValidator,
+   ERR_INVALID_STEP_CONFIG,
+   ERR_UNSUPPORTED_STEP,
+} from "@mate/sdk";
 import { isLeft, isRight } from "fp-ts/lib/Either";
 import { PathReporter } from "io-ts/lib/PathReporter";
 import { ValidationError } from "./ValidationError";
@@ -14,10 +18,8 @@ import { IExchangeServiceRepo } from "../exchanges";
 import {
    ERR_ASSET_UNSUPPORTED_ON_EXCHANGE,
    ERR_EXCHANGE_SERVICE_NOT_FOUND,
-   ERR_INVALID_STEP_CONFIG,
    ERR_UNSUPPORTED_CHAIN,
    ERR_UNSUPPORTED_EXCHANGE,
-   ERR_UNSUPPORTED_STEP,
 } from "../constants";
 
 @injectable()
