+++ packages/sdk/src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from "./ConfigSchema";
+export * from "./ErrorMessages";
